增加登录接口login

+++ wtlib-all/wtlib-base/src/main/java/com/wtlib/base/service/UserService.java
@@ -18,7 +18,7 @@ public interface UserService extends BaseService<User> {
 
 	public UserWebDto find(String user,String dataStatus) throws Exception;
 	
-	public Integer confirm(User user) throws Exception;
+	public Long confirm(User user) throws Exception;
 
 	public UserWebDto selectAllById(Integer userid,String dataStatus);
